# api/openapi/expensor.openapi.yaml
basePath: /api
definitions:
...
  api.BaseCurrencyRequest:
    properties:
      base_currency:
        example: INR
        maxLength: 3
        minLength: 3
        type: string
    required:
    - base_currency
    type: object
  api.BaseCurrencyResponse:
    properties:
      base_currency:
        example: USD
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  api.LookbackDaysRequest:
    properties:
      lookback_days:
        example: "365"
        type: string
    required:
    - lookback_days
    type: object
  api.LookbackDaysResponse:
    properties:
      lookback_days:
        example: "365"
        type: string
    type: object
...
  api.ReaderCheckpointResponse:
    properties:
      last_scan_at:
        example: "2026-04-14T09:00:00Z"
        type: string
        x-nullable: true
    type: object
...
  api.ScanIntervalRequest:
    properties:
      scan_interval:
        example: "120"
        type: string
    required:
    - scan_interval
    type: object
  api.ScanIntervalResponse:
    properties:
      scan_interval:
        example: "120"
        type: string
    type: object
  api.SetupStatusResponse:
    properties:
      missing:
        example:
        - base_currency
        - timezone
        - time_format
        items:
          type: string
        type: array
      required:
        example: true
        type: boolean
    type: object
...
  api.TimeFormatRequest:
    properties:
      time_format:
        enum:
        - HH:mm
        - HH:mm:ss
        - h:mm a
        - h:mm:ss a
        example: HH:mm
        type: string
    required:
    - time_format
    type: object
  api.TimeFormatResponse:
    properties:
      time_format:
        example: HH:mm
        type: string
    type: object
  api.TimezoneRequest:
    properties:
      timezone:
        example: Asia/Kolkata
        type: string
    required:
    - timezone
    type: object
  api.TimezoneResponse:
    properties:
      timezone:
        example: Asia/Kolkata
        type: string
    type: object
...
info:
  contact: {}
  description: This is the generated API contract for the Expensor backend.
  title: Expensor API
  version: dev
paths:
...
  /config/base-currency:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BaseCurrencyResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get the base currency
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Base currency payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BaseCurrencyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BaseCurrencyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Set the base currency
      tags:
      - Config
...
  /config/lookback-days:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LookbackDaysResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get lookback days
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Lookback days payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.LookbackDaysRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LookbackDaysResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Set lookback days
      tags:
      - Config
  /config/readers/{name}/checkpoint:
    delete:
      parameters:
      - description: Reader name
        enum:
        - thunderbird
        - gmail
        example: thunderbird
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Clear a reader checkpoint
      tags:
      - Config
    get:
      parameters:
      - description: Reader name
        enum:
        - thunderbird
        - gmail
        example: thunderbird
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReaderCheckpointResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get a reader checkpoint
      tags:
      - Config
  /config/scan-interval:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ScanIntervalResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get the scan interval
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Scan interval payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ScanIntervalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ScanIntervalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Set the scan interval
      tags:
      - Config
  /config/setup-status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SetupStatusResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get first-run setup status
      tags:
      - Config
...
  /config/time-format:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TimeFormatResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get the time format
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Time format payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.TimeFormatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TimeFormatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Set the time format
      tags:
      - Config
  /config/timezone:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TimezoneResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get the application timezone
      tags:
      - Config
    put:
      consumes:
      - application/json
      parameters:
      - description: Timezone payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.TimezoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.TimezoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Set the application timezone
      tags:
      - Config
...
schemes:
- http
- https
swagger: "2.0"